requirements.txt
absl-py==1.0.0
altgraph==0.17.2
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
cloudpickle==2.0.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
et-xmlfile==1.1.0
flatbuffers==2.0
fonttools @ file:///D:/bld/fonttools_1651018050865/work
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gym==0.21.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///D:/bld/kiwisolver_1648854623027/work
libclang==12.0.0
lxml==4.8.0
macholib==1.16
Markdown==3.3.6
matplotlib @ file:///D:/bld/matplotlib-suite_1639359207070/work
munkres==1.1.4
networkx==2.6.3
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openpyxl==3.0.9
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas @ file:///D:/bld/pandas_1639398349358/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1632667180946/work
pefile==2021.9.3
Pillow @ file:///D:/bld/pillow_1649818245820/work
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQtChart==5.15.5
PyQtChart-Qt5==5.15.2
pyqtgraph==0.12.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-docx==0.8.11
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
pywin32==303
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-fuzzy==0.4.2
scikit-learn==1.0.2
scipy @ file:///C:/bld/scipy_1637806996411/work
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1629095986539/work
shiboken2==5.15.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
statsmodels @ file:///D:/bld/statsmodels_1644535845397/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650370875435/work
unicodedata2 @ file:///D:/bld/unicodedata2_1649112131705/work
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
